Multislice & object comparisons (#62)
- Add support for even/odd split reconstructions
- Support object comparisons and Fourier Ring Correlation (FRC)
- Support loading and viewing multi-layer (multislice) objects including layer distance
- Relocate diffraction pattern loading to wizard in detectors view
- Simplify behavior of diffraction pattern crop widgets
- Add new mouse tools to image view: move, ruler, rectangle, line-cut
- Add option to view complex array intensity
- Add color legends for acyclic and cyclic quantities to image view
- Fix tike cost function plots
- Add memory usage monitor widget
- Add Fourier Zone Plate (FZP) presets for several Advanced Photon Source (APS) instruments
- Add space-deliminated scan position reader to support format used for NXSchool IC datasets
- Update interface to support PtychoNN v0.2
- Extract probe/object classes and refactor

+++ ptychodus/api/data.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from dataclasses import dataclass
-from decimal import Decimal
 from enum import Enum, auto
 from pathlib import Path
 from typing import overload, Any, Optional, TypeAlias, Union
@@ -83,12 +82,12 @@ class DiffractionMetadata:
     numberOfPatternsPerArray: int
     numberOfPatternsTotal: int
     patternDataType: numpy.dtype[numpy.integer[Any]]
-    detectorDistanceInMeters: Optional[Decimal] = None
+    detectorDistanceInMeters: Optional[float] = None
     detectorExtentInPixels: Optional[ImageExtent] = None
     detectorPixelGeometry: Optional[PixelGeometry] = None
     detectorBitDepth: Optional[int] = None
     cropCenterInPixels: Optional[Array2D[int]] = None
-    probeEnergyInElectronVolts: Optional[Decimal] = None
+    probeEnergyInElectronVolts: Optional[float] = None
     filePath: Optional[Path] = None
 
     @classmethod
